+++ app/src/main/java/by/app/slise/MovieDetails/Actors.kt
@@ -1,4 +1,4 @@
-package by.app.slise.ActorsRecyclerView
+package by.app.slise.MovieDetails
 
 data class Actors(
         val actorView: String,
