+++ app/src/main/java/by/app/slise/MovieDetails/ActorsDatasource.kt
@@ -1,4 +1,4 @@
-package by.app.slise.ActorsRecyclerView
+package by.app.slise.MovieDetails
 
 
 class ActorsDatasource {
