+++ app/src/main/java/by/app/slise/MovieList/Movies.kt
@@ -1,4 +1,4 @@
-package by.app.slise.MoviesRecyclerView
+package by.app.slise.MovieList
 
 
 data class Movies(
